# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tracker API",
    "version": "1.0.0",
    "description": "REST API of the tracker. Humans and AI agents use the same API; the `tracker` CLI and the web app are clients of it.\n\n**Conventions**\n- Authenticate with `Authorization: Bearer trk_…` (API token) or the web session cookie.\n- Paths accept human references: issues as `ENG-42` or id, projects as `ENG` or id, users as `@handle`, `me` or id.\n- Lists return `{ data, nextCursor }`; pass `cursor` to continue.\n- Errors are RFC 9457 `application/problem+json` with a stable `code`.\n- Issues carry `version`; send `If-Match: \"v<version>\"` to avoid lost updates (412 `VERSION_MISMATCH`).\n- POST requests accept `Idempotency-Key` for safe retries.\n- Every change is recorded in the event log (`GET /events`) and delivered to webhooks."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "sessionCookie": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API token (`trk_…`), created with `POST /users/{user}/tokens` or `poietic-issues token create`."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "poietic_issues_session",
        "description": "Web session from `/auth/token-login` or `/auth/dev-login`."
      }
    },
    "schemas": {
      "Event": {
        "type": "object",
        "properties": {
          "seq": {
            "type": "integer",
            "description": "Strictly increasing in commit order (may skip values). Use as a cursor."
          },
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "issue.created",
              "issue.updated",
              "issue.deleted",
              "issue.restored",
              "comment.created",
              "comment.updated",
              "comment.deleted",
              "link.created",
              "link.deleted",
              "project.created",
              "project.updated",
              "status.created",
              "status.updated",
              "status.deleted",
              "label.created",
              "label.updated",
              "label.deleted",
              "user.created",
              "user.updated",
              "field.created",
              "field.updated",
              "field.deleted",
              "attachment.created",
              "attachment.deleted"
            ]
          },
          "actorId": {
            "type": [
              "string",
              "null"
            ]
          },
          "actor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "projectId": {
            "type": [
              "string",
              "null"
            ]
          },
          "issueId": {
            "type": [
              "string",
              "null"
            ]
          },
          "data": {
            "type": "object",
            "additionalProperties": {},
            "description": "Payload; shape depends on `type`."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "ISO-8601 UTC timestamp.",
            "example": "2026-09-26T13:49:00.123Z"
          }
        },
        "required": [
          "seq",
          "id",
          "type",
          "actorId",
          "actor",
          "projectId",
          "issueId",
          "data",
          "createdAt"
        ]
      },
      "UserSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "usr_01h455vb4pex5vsknk084sn02q"
          },
          "handle": {
            "type": "string",
            "example": "ada"
          },
          "name": {
            "type": "string",
            "example": "Ada Lovelace"
          },
          "kind": {
            "type": "string",
            "enum": [
              "human",
              "agent",
              "system"
            ],
            "description": "Humans and AI agents are both first-class actors; `system` is internal."
          },
          "avatarUrl": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "handle",
          "name",
          "kind",
          "avatarUrl"
        ]
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "usr_01h455vb4pex5vsknk084sn02q"
          },
          "handle": {
            "type": "string",
            "example": "ada"
          },
          "name": {
            "type": "string",
            "example": "Ada Lovelace"
          },
          "kind": {
            "type": "string",
            "enum": [
              "human",
              "agent",
              "system"
            ],
            "description": "Humans and AI agents are both first-class actors; `system` is internal."
          },
          "avatarUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "member"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "ISO-8601 UTC timestamp.",
            "example": "2026-09-26T13:49:00.123Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "ISO-8601 UTC timestamp.",
            "example": "2026-09-26T13:49:00.123Z"
          },
          "deactivatedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "ISO-8601 UTC timestamp.",
            "example": "2026-09-26T13:49:00.123Z"
          }
        },
        "required": [
          "id",
          "handle",
          "name",
          "kind",
          "avatarUrl",
          "email",
          "role",
          "createdAt",
          "updatedAt",
          "deactivatedAt"
        ]
      },
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "urn:tracker:error:NOT_FOUND"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_FAILED",
              "UNAUTHENTICATED",
              "FORBIDDEN",
              "PENDING_APPROVAL",
              "NOT_FOUND",
              "CONFLICT",
              "IDEMPOTENCY_IN_PROGRESS",
              "VERSION_MISMATCH",
              "PAYLOAD_TOO_LARGE",
              "UNSUPPORTED_MEDIA_TYPE",
              "INVALID_RELATION",
              "IDEMPOTENCY_KEY_REUSED",
              "RATE_LIMITED",
              "INTERNAL",
              "UNAVAILABLE"
            ]
          },
          "detail": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string",
                  "description": "Dotted path of the offending input field.",
                  "example": "title"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ]
            }
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "description": "RFC 9457 problem details with a stable machine-readable `code`."
      },
      "TokenLoginInput": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "required": [
          "token"
        ]
      },
      "DevLoginInput": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "description": "User id or handle.",
            "example": "ada"
          }
        },
        "required": [
          "user"
        ]
      },
...
      "ApiToken": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "prefix": {
            "type": "string",
            "description": "First characters of the token, for recognition.",
            "example": "trk_a1b2"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "ISO-8601 UTC timestamp.",
            "example": "2026-09-26T13:49:00.123Z"
          },
          "lastUsedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "ISO-8601 UTC timestamp.",
            "example": "2026-09-26T13:49:00.123Z"
          },
          "expiresAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "ISO-8601 UTC timestamp.",
            "example": "2026-09-26T13:49:00.123Z"
          },
          "revokedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "ISO-8601 UTC timestamp.",
            "example": "2026-09-26T13:49:00.123Z"
          }
        },
        "required": [
          "id",
          "userId",
          "name",
          "prefix",
          "scopes",
          "createdAt",
          "lastUsedAt",
          "expiresAt",
          "revokedAt"
        ]
      },
      "CreatedApiToken": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "prefix": {
            "type": "string",
            "description": "First characters of the token, for recognition.",
            "example": "trk_a1b2"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "ISO-8601 UTC timestamp.",
            "example": "2026-09-26T13:49:00.123Z"
          },
          "lastUsedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "ISO-8601 UTC timestamp.",
            "example": "2026-09-26T13:49:00.123Z"
          },
          "expiresAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "ISO-8601 UTC timestamp.",
            "example": "2026-09-26T13:49:00.123Z"
          },
          "revokedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "ISO-8601 UTC timestamp.",
            "example": "2026-09-26T13:49:00.123Z"
          },
          "token": {
            "type": "string",
            "description": "The secret token. Shown only once — store it now."
          }
        },
        "required": [
          "id",
          "userId",
          "name",
          "prefix",
          "scopes",
          "createdAt",
          "lastUsedAt",
          "expiresAt",
          "revokedAt",
          "token"
        ]
      },
      "CreateTokenInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "expiresAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "ISO-8601 UTC timestamp.",
            "example": "2026-09-26T13:49:00.123Z"
          }
        },
        "required": [
          "name"
        ]
      },
...
      "EventPage": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Event"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Opaque cursor for the next page; null when this is the last page."
          }
        },
        "required": [
          "data",
          "nextCursor"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/auth/token-login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "security": [],
        "summary": "Sign in with an API token",
        "description": "Exchanges an API token for an HttpOnly session cookie (used by the web UI).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenLoginInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Problem details. Codes: VALIDATION_FAILED.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Problem details. Codes: UNAUTHENTICATED.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/auth/dev-login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "security": [],
        "summary": "Sign in as any user (development only)",
        "description": "Only available when the server runs with `POIETIC_ISSUES_AUTH_MODE=dev`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevLoginInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Problem details. Codes: VALIDATION_FAILED.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Problem details. Codes: UNAUTHENTICATED.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Problem details. Codes: FORBIDDEN.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Problem details. Codes: NOT_FOUND.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/{user}/tokens": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List a user's API tokens",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "me"
            },
            "required": true,
            "name": "user",
            "in": "path",
            "description": "User id, handle, `@handle` or `me`."
          }
        ],
        "responses": {
          "200": {
            "description": "Tokens",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ApiToken"
                      }
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Problem details. Codes: VALIDATION_FAILED.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Problem details. Codes: UNAUTHENTICATED.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Problem details. Codes: FORBIDDEN.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Problem details. Codes: NOT_FOUND.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create an API token",
        "description": "The secret `token` is returned only in this response.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "me"
            },
            "required": true,
            "name": "user",
            "in": "path",
            "description": "User id, handle, `@handle` or `me`."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTokenInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedApiToken"
                }
              }
            }
          },
          "400": {
            "description": "Problem details. Codes: VALIDATION_FAILED.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Problem details. Codes: UNAUTHENTICATED.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Problem details. Codes: FORBIDDEN.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Problem details. Codes: NOT_FOUND.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/events": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Read the event log",
        "description": "Events in commit order. Poll with `after=<last seq seen>`; `nextCursor` is the seq to pass next. Ideal for agents that need \"what changed since X\".",
        "parameters": [
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            },
            "required": false,
            "name": "after",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Project key or id."
            },
            "required": false,
            "description": "Project key or id.",
            "name": "project",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Issue key or id."
            },
            "required": false,
            "description": "Issue key or id.",
            "name": "issue",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Comma-separated event types: issue.created, issue.updated, issue.deleted, issue.restored, comment.created, comment.updated, comment.deleted, link.created, link.deleted, project.created, project.updated, status.created, status.updated, status.deleted, label.created, label.updated, label.deleted, user.created, user.updated, field.created, field.updated, field.deleted, attachment.created, attachment.deleted."
            },
            "required": false,
            "description": "Comma-separated event types: issue.created, issue.updated, issue.deleted, issue.restored, comment.created, comment.updated, comment.deleted, link.created, link.deleted, project.created, project.updated, status.created, status.updated, status.deleted, label.created, label.updated, label.deleted, user.created, user.updated, field.created, field.updated, field.deleted, attachment.created, attachment.deleted.",
            "name": "types",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Events",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPage"
                }
              }
            }
          },
          "400": {
            "description": "Problem details. Codes: VALIDATION_FAILED.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Problem details. Codes: UNAUTHENTICATED.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Problem details. Codes: NOT_FOUND.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}